更新webui，可以删除消息了

+++ service-memory/src/config.rs
@@ -2,7 +2,7 @@ use serde::Deserialize;
 
 #[derive(Debug, Deserialize)]
 pub struct MemoryConfig {
-    pub short_term_length: usize,
+    pub short_term_length: i64,
     pub storage_dir: String,
-    pub display_length: usize,
+    pub display_length: i64,
 }
